+++ MQTTClient/MQTTClient/MQTTTransport.h
@@ -45,8 +45,8 @@
 /** host an NSString containing the hostName or IP address of the host to connect to */
 @property (strong, nonatomic) NSString * _Nonnull host;
 
-/** port an unsigned 16 bit integer containing the IP port number to connect to */
-@property (nonatomic) UInt16 port;
+/** port an unsigned 32 bit integer containing the IP port number to connect to */
+@property (nonatomic) UInt32 port;
 
 /** MQTTTransportDelegate needs to be set to a class implementing th MQTTTransportDelegate protocol
  * to receive delegate messages.
